+++ gridy/Sources/ContentView.swift
@@ -11,8 +11,9 @@ import ComposableArchitecture
 struct ContentView: View {
     var body: some View {
         ZStack {
-            BackgroundView()
-            AuthenticationView()
+            TimelineLayoutView()
+//            BackgroundView()
+//            AuthenticationView()
         }
     }
 }
